Move PropertyExtractor to base services
