+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/properties/TypePropertyMetadata.java
@@ -17,6 +17,7 @@
 package org.gradle.api.internal.tasks.properties;
 
 import org.gradle.api.internal.tasks.properties.annotations.PropertyAnnotationHandler;
+import org.gradle.internal.reflect.PropertyMetadata;
 
 import javax.annotation.Nullable;
 import java.util.Set;
