+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/properties/DefaultTypePropertyMetadataStoreTest.groovy
@@ -40,6 +40,7 @@ import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.OutputFiles
 import org.gradle.cache.internal.TestCrossBuildInMemoryCacheFactory
+import org.gradle.internal.reflect.PropertyMetadata
 import org.gradle.internal.scripts.ScriptOrigin
 import spock.lang.Issue
 import spock.lang.Shared
